pyproject.toml
[tool.poetry]
name = "settle-oauth-gcp"
version = "0.1.0"
description = ""
authors = ["Settle <support@settle.eu>"]

[tool.poetry.dependencies]
python = "^3.9"
google-auth = "^2.0.1"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
